format and testing library
