+++ google/cloud/pubsublite/CMakeLists.txt
@@ -184,9 +184,6 @@ add_library(google_cloud_cpp_pubsublite_mocks INTERFACE)
 target_sources(
     google_cloud_cpp_pubsublite_mocks
     INTERFACE ${CMAKE_CURRENT_SOURCE_DIR}/mocks/mock_admin_connection.h
-              ${CMAKE_CURRENT_SOURCE_DIR}/internal/mock_async_reader_writer.h
-              ${CMAKE_CURRENT_SOURCE_DIR}/internal/mock_backoff_policy.h
-              ${CMAKE_CURRENT_SOURCE_DIR}/internal/mock_retry_policy.h
               ${CMAKE_CURRENT_SOURCE_DIR}/mocks/mock_topic_stats_connection.h)
 target_link_libraries(
     google_cloud_cpp_pubsublite_mocks
@@ -211,6 +208,20 @@ function (google_cloud_cpp_pubsublite_client_define_tests)
     # googletest itself), because the generic `FindGTest` module does not define
     # the GTest::gmock target, and the target names are also weird.
     find_package(GTest CONFIG REQUIRED)
+
+    add_library(
+        pubsublite_testing # cmake-format: sort
+        testing/mock_async_reader_writer.h testing/mock_backoff_policy.h
+        testing/mock_retry_policy.h)
+    target_link_libraries(
+        pubsublite_testing
+        PUBLIC google-cloud-cpp::pubsublite google-cloud-cpp::grpc_utils
+               google-cloud-cpp::common google-cloud-cpp::pubsublite_protos
+               GTest::gmock GTest::gtest)
+    google_cloud_cpp_add_common_options(pubsublite_testing)
+    target_compile_options(pubsublite_testing
+                           PUBLIC ${GOOGLE_CLOUD_CPP_EXCEPTIONS_FLAG})
+    create_bazel_config(pubsublite_testing YEAR "2022")
 
     set(pubsublite_unit_tests # cmake-format: sort
                               endpoint_test.cc internal/stream_factory_test.cc)
