+++ google/cloud/pubsublite/pubsublite_testing.bzl
@@ -1,4 +1,4 @@
-# Copyright 2021 Google LLC
+# Copyright 2020 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -14,12 +14,10 @@
 #
 # DO NOT EDIT -- GENERATED BY CMake -- Change the CMakeLists.txt file if needed
 
-"""Automatically generated source lists for google_cloud_cpp_pubsublite_mocks - DO NOT EDIT."""
+"""Automatically generated source lists for pubsublite_testing - DO NOT EDIT."""
 
-google_cloud_cpp_pubsublite_mocks_hdrs = [
-    "mocks/mock_admin_connection.h",
-    "mocks/mock_topic_stats_connection.h",
-]
-
-google_cloud_cpp_pubsublite_mocks_srcs = [
+pubsublite_testing_hdrs = [
+    "testing/mock_async_reader_writer.h",
+    "testing/mock_backoff_policy.h",
+    "testing/mock_retry_policy.h",
 ]
