+++ google/cloud/pubsublite/testing/mock_retry_policy.h
@@ -12,27 +12,28 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_PUBSUBLITE_INTERNAL_MOCK_BACKOFF_POLICY_H
-#define GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_PUBSUBLITE_INTERNAL_MOCK_BACKOFF_POLICY_H
+#ifndef GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_PUBSUBLITE_TESTING_MOCK_RETRY_POLICY_H
+#define GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_PUBSUBLITE_TESTING_MOCK_RETRY_POLICY_H
 
-#include "google/cloud/backoff_policy.h"
+#include "google/cloud/internal/retry_policy.h"
 #include "google/cloud/version.h"
 #include <gmock/gmock.h>
 
 namespace google {
 namespace cloud {
 GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_BEGIN
-namespace pubsublite_internal {
+namespace pubsublite_testing {
 
-class MockBackoffPolicy : public BackoffPolicy {
+class MockRetryPolicy : public RetryPolicy {
  public:
-  MOCK_METHOD(std::unique_ptr<BackoffPolicy>, clone, (), (const override));
-  MOCK_METHOD(std::chrono::milliseconds, OnCompletion, (), (override));
+  MOCK_METHOD(bool, OnFailure, (Status const&), (override));
+  MOCK_METHOD(bool, IsExhausted, (), (const, override));
+  MOCK_METHOD(bool, IsPermanentFailure, (Status const&), (const, override));
 };
 
-}  // namespace pubsublite_internal
+}  // namespace pubsublite_testing
 GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_END
 }  // namespace cloud
 }  // namespace google
 
-#endif  // GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_PUBSUBLITE_INTERNAL_MOCK_BACKOFF_POLICY_H
+#endif  // GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_PUBSUBLITE_TESTING_MOCK_RETRY_POLICY_H
